chore(interface): rename ApiKeyModal
